+++ R/export_data.r
@@ -124,11 +124,10 @@ export_control_global <- function(control = NULL) {
 #' @export
 
 export_object <- function(
-  keyword = NULL,
-  object = NULL,
-  control = NULL,
-  location = NULL
-) {
+    keyword = NULL,
+    object = NULL,
+    control = NULL,
+    location = NULL) {
   out <- .export_data(
     table = gt.env$tbl_object,
     in_keyword = unlist(keyword),
@@ -145,10 +144,9 @@ export_object <- function(
 #' @export
 
 export_object_global <- function(
-  keyword = NULL,
-  object = NULL,
-  control = NULL
-) {
+    keyword = NULL,
+    object = NULL,
+    control = NULL) {
   out <- .export_data(
     table = gt.env$tbl_object,
     in_keyword = unlist(keyword),
@@ -164,11 +162,10 @@ export_object_global <- function(
 #' @export
 
 export_score <- function(
-  keyword = NULL,
-  object = NULL,
-  control = NULL,
-  location = NULL
-) {
+    keyword = NULL,
+    object = NULL,
+    control = NULL,
+    location = NULL) {
   out <- .export_data(
     table = gt.env$tbl_score,
     in_keyword = unlist(keyword),
@@ -176,9 +173,8 @@ export_score <- function(
     in_control = unlist(control),
     in_location = unlist(location),
   )
-  out <- filter(out, .data$location != "world" & .data$synonym == 0)
+  out <- filter(out, .data$location != "world")
   out <- rename(out, control = batch_c, object = batch_o)
-  out <- select(out, -synonym)
   return(out)
 }
 
@@ -193,9 +189,7 @@ export_voi <- function(keyword = NULL, object = NULL, control = NULL) {
     in_control = unlist(control),
     in_location = "world"
   )
-  out <- filter(out, .data$synonym == 0)
   out <- rename(out, control = batch_c, object = batch_o)
-  out <- select(out, -synonym)
   return(out)
 }
 
@@ -203,11 +197,10 @@ export_voi <- function(keyword = NULL, object = NULL, control = NULL) {
 #' @export
 
 export_doi <- function(
-  keyword = NULL,
-  object = NULL,
-  control = NULL,
-  locations = NULL
-) {
+    keyword = NULL,
+    object = NULL,
+    control = NULL,
+    locations = NULL) {
   out <- .export_data(
     table = gt.env$tbl_doi,
     in_keyword = unlist(keyword),
@@ -230,14 +223,13 @@ export_doi <- function(
 #' @importFrom lubridate as_date
 
 .export_data <- function(
-  table,
-  in_keyword = NULL,
-  in_object = NULL,
-  in_control = NULL,
-  in_batch = NULL,
-  in_location = NULL,
-  in_locations = NULL
-) {
+    table,
+    in_keyword = NULL,
+    in_object = NULL,
+    in_control = NULL,
+    in_batch = NULL,
+    in_location = NULL,
+    in_locations = NULL) {
   keyword <- in_keyword
   object <- in_object
   control <- in_control
@@ -284,6 +276,6 @@ export_doi <- function(
   }
 
   table <- collect(table)
-  table <- mutate(table, date = as_date(paste0(.data$date, "-01")))
+  table <- mutate(table, date = as_date(.data$date))
   return(table)
 }